Move meme logic out of aurebesh
